requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=1.0.0=pyhd8ed1ab_0
alabaster=0.7.12=py_0
antlr4-python3-runtime=4.8=pypi_0
anyjson=0.3.3=pyhd8ed1ab_1004
argcomplete=1.12.3=pypi_0
argon2-cffi=21.1.0=py37h5e8e339_2
arrow=1.2.1=pyhd8ed1ab_0
astor=0.8.1=pyh9f0ad1d_0
attrs=21.2.0=pyhd8ed1ab_0
autoflake=1.3=pyhd8ed1ab_1
autopep8=1.5.6=pyhd8ed1ab_1
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=pypi_0
backports-entry-points-selectable=1.1.1=pypi_0
bcrypt=3.2.0=py37h5e8e339_2
blas=2.112=mkl
blas-devel=3.9.0=12_linux64_mkl
bleach=4.1.0=pypi_0
brotlipy=0.7.0=py37h5e8e339_1003
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f98852_0
ca-certificates=2021.10.8=ha878542_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cairo=1.16.0=hcf35c78_1003
certifi=2021.10.8=py37h89c1867_1
cffi=1.14.4=py37h11fe52a_0
cfgv=3.3.1=pypi_0
chardet=4.0.0=py37h89c1867_2
colour=0.1.5=py_0
coverage=5.3.1=py37h5e8e339_1
coveralls=3.0.1=pyhd8ed1ab_0
cryptography=36.0.0=py37hf1a17b8_0
cudatoolkit=10.2.89=h8f6ccaa_9
cycler=0.11.0=pypi_0
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.6=hfdff14a_1
debugpy=1.5.1=pypi_0
decorator=5.1.0=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.4=pypi_0
diskcache=5.2.1=pypi_0
distlib=0.3.3=pypi_0
docopt=0.6.2=py_1
docutils=0.16=py37h89c1867_3
entrypoints=0.3=pypi_0
expat=2.4.1=h9c3ff4c_0
facenet-pytorch=2.5.2=pypi_0
ffmpeg=4.4.0=hca11adc_2
filelock=3.4.0=pypi_0
flake8=3.9.0=pyhd8ed1ab_0
fontconfig=2.13.1=hba837de_1005
fonttools=4.28.2=pypi_0
freetype=2.10.4=h0708190_1
furl=2.1.3=pyhd8ed1ab_0
future=0.18.2=pypi_0
gast=0.5.3=pyhd8ed1ab_0
gettext=0.19.8.1=hf34092f_1004
glib=2.66.3=h58526e2_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-pasta=0.2.0=pyh8c360ce_0
graphite2=1.3.13=h58526e2_1001
grpcio=1.42.0=py37hb27c1af_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
h5py=3.3.0=nompi_py37ha3df211_100
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.6=nompi_h6a2412b_1114
icu=64.2=he1b5a44_1
identify=2.4.0=pypi_0
idna=2.10=pyh9f0ad1d_0
imageio=2.12.0=pypi_0
imagesize=1.3.0=pyhd8ed1ab_0
importlib-metadata=4.8.2=py37h89c1867_0
importlib-resources=5.4.0=pypi_0
importlib_metadata=4.8.2=hd8ed1ab_0
infinity=1.5=pyhd8ed1ab_0
iniconfig=1.1.1=pyh9f0ad1d_0
intervals=0.9.2=pyhd8ed1ab_0
ipykernel=6.5.1=pypi_0
ipython=7.29.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jasper=1.900.1=h07fcdf6_1006
jbig=2.1=h7f98852_2003
jedi=0.18.1=pypi_0
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
jsonschema=4.2.1=pypi_0
jupyter-client=7.1.0=pypi_0
jupyter-core=4.9.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.2=pypi_0
krb5=1.19.2=hcc1bbae_3
lame=3.100=h7f98852_1001
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libclang=9.0.1=default_ha53f305_2
libcurl=7.80.0=h2574ce0_0
libdeflate=1.8=h7f98852_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.2.1=he1b5a44_1007
libgcc-ng=11.2.0=h1d223b6_11
libgfortran-ng=11.2.0=h69a702a_11
libgfortran5=11.2.0=h5c6108e_11
libglib=2.66.3=hbe7bbb4_0
libiconv=1.16=h516909a_0
liblapack=3.9.0=12_linux64_mkl
liblapacke=3.9.0=12_linux64_mkl
libllvm9=9.0.1=default_hc23dcda_4
libnghttp2=1.43.0=h812cca2_1
libopencv=4.5.1=py37h5fff631_1
libpng=1.6.37=h21135ba_2
libprotobuf=3.15.8=h780b84a_1
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=he4da1e4_11
libtiff=4.3.0=h6f004c6_2
libuuid=2.32.1=h7f98852_1000
libvpx=1.11.0=h9c3ff4c_3
libwebp-base=1.2.1=h7f98852_0
libxcb=1.13=h7f98852_1004
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=hee79883_0
libzlib=1.2.11=h36c2ea0_1013
llvm-openmp=12.0.1=h4bd325d_1
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.6=pyhd8ed1ab_0
markupsafe=2.0.1=py37h5e8e339_1
matplotlib=3.5.0=pypi_0
matplotlib-inline=0.1.3=pypi_0
mccabe=0.6.1=py_1
mistune=0.8.4=pypi_0
mkl=2021.4.0=h8d4b97c_729
mkl-devel=2021.4.0=ha770c72_730
mkl-include=2021.4.0=h8d4b97c_729
mock=4.0.3=py37h89c1867_2
more-itertools=8.11.0=pyhd8ed1ab_0
nbclient=0.5.9=pypi_0
nbconvert=6.3.0=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=h58526e2_4
nest-asyncio=1.5.1=pypi_0
nettle=3.6=he412f7d_0
networkx=2.6.3=pypi_0
ninja=1.10.2=h4bd325d_1
nodeenv=1.6.0=pypi_0
notebook=6.4.6=pypi_0
nspr=4.32=h9c3ff4c_1
nss=3.72=hb5efdd6_0
numpy=1.20.1=py37haa41c4c_0
olefile=0.46=pyh9f0ad1d_1
opencv=4.5.1=py37h89c1867_1
openh264=2.1.1=h780b84a_0
openssl=1.1.1l=h7f98852_0
orderedmultidict=1.0.1=py_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.2.3=py37hdc94413_0
pandocfilters=1.5.0=pypi_0
parso=0.8.2=pypi_0
passlib=1.7.4=pyh9f0ad1d_0
pcre=8.45=h9c3ff4c_0
petastorm=0.9.8=pypi_0
pexpect=4.8.0=pypi_0
phonenumbers=8.12.37=pyhd8ed1ab_0
pickleshare=0.7.5=pypi_0
pillow=8.4.0=pypi_0
pip=21.0.1=pyhd8ed1ab_0
pixman=0.38.0=h516909a_1003
platformdirs=2.4.0=pypi_0
pluggy=0.13.1=pypi_0
pre-commit=2.8.2=pypi_0
prometheus-client=0.12.0=pypi_0
prompt-toolkit=3.0.22=pypi_0
protobuf=3.15.8=py37hcd2ae1e_0
psutil=5.8.0=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pypi_0
py=1.11.0=pyh6c4a22f_0
py-opencv=4.5.1=py37h085eea5_1
py4j=0.10.9=pypi_0
pyarrow=6.0.1=pypi_0
pycodestyle=2.7.0=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pyflakes=2.3.1=pyhd8ed1ab_0
pygments=2.10.0=pyhd8ed1ab_0
pymysql=0.10.1=pyhd3deb0d_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=3.0.6=pyhd8ed1ab_0
pyrsistent=0.18.0=pypi_0
pysocks=1.7.1=py37h89c1867_4
pyspark=3.0.2=pypi_0
pytest=6.1.2=pypi_0
pytest-asyncio=0.14.0=py37h89c1867_0
pytest-cov=2.11.1=pyh44b312d_0
python=3.7.6=cpython_h8356626_6
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytorch=1.6.0=py3.7_cuda10.2.89_cudnn7.6.5_0
pytz=2021.3=pyhd8ed1ab_0
pywavelets=1.2.0=pypi_0
pyyaml=6.0=py37h5e8e339_3
pyzmq=22.3.0=pypi_0
qt=5.12.5=hd8c4c69_1
readline=8.1=h46c0cb4_0
requests=2.25.1=pyhd3deb0d_0
scikit-image=0.18.3=pypi_0
scikit-learn=0.23.2=py37hddcf8d6_3
scipy=1.7.2=py37hf2a6cf1_0
send2trash=1.8.0=pypi_0
setuptools=59.2.0=py37h89c1867_0
setuptools-scm=6.3.2=pypi_0
six=1.16.0=pyh6c4a22f_0
sklearn=0.0=pypi_0
snowballstemmer=2.2.0=pyhd8ed1ab_0
sphinx=3.3.0=pyhd8ed1ab_0
sphinx_rtd_theme=0.5.0=pyh9f0ad1d_0
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=2.0.0=pyhd8ed1ab_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.5=pyhd8ed1ab_1
sqlalchemy=1.3.20=py37h5e8e339_0
sqlalchemy-utils=0.36.6=pyh9f0ad1d_0
sqlite=3.36.0=h9cd32fc_2
tbb=2021.4.0=h4bd325d_1
tensorboard=1.14.0=py37_0
tensorflow=1.14.0=h4531e10_0
tensorflow-base=1.14.0=py37h4531e10_0
tensorflow-estimator=1.14.0=py37h5ca1d4c_0
termcolor=1.1.0=py_2
terminado=0.12.1=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=3.0.0=pyh8a188c0_0
tifffile=2021.11.2=pypi_0
tk=8.6.11=h27826a3_1
toml=0.10.2=pyhd8ed1ab_0
tomli=1.2.2=pypi_0
torchvision=0.7.0=py37_cu102
tornado=6.1=pypi_0
traitlets=5.1.1=pypi_0
typing_extensions=4.0.0=pyha770c72_0
urllib3=1.26.7=pyhd8ed1ab_0
virtualenv=20.10.0=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.37.0=pyhd8ed1ab_1
widgetsnbextension=3.5.2=pypi_0
wrapt=1.13.3=py37h5e8e339_1
x264=1!161.3030=h7f98852_1
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.11=h36c2ea0_1013
zstd=1.5.0=ha95c52a_0
